Add Dusk test to confirm .env.dusk.local is not working
Bug Description
===============

In the Laravel Dusk documentation is states that you can create a
seperate ENV file for Dusk testing by naming it .env.dusk.{environment}.

When I do this it appears that the ENV values that were set in .env.local
are still used by the system and the dusk specific file is ignored.

This may be something to do with the combination of SAIL + Dusk?

Reproduction
============

1. Start the sail app:

```
./vendor/bin/sail up
```

2. Open the `.env` file in your IDE so you can watch it change as the ENV
   config is updated by Dusk.

3. Run e2e test
```
./vendor/bin/sail dusk
```

4. See that:
  * 1 test fails in `tests/Browser/EnvTest`
  * The first test in the same test file passes when it shouldn't